chore(test-tags): reorganize test_tags folder to match UI folder structure
