+++ app/src/main/java/com/android/unio/model/strings/test_tags/navigation/NavigationTestTags.kt
@@ -1,4 +1,4 @@
-package com.android.unio.model.strings.test_tags
+package com.android.unio.model.strings.test_tags.navigation
 
 object NavigationActionTestTags {
   const val BOTTOM_NAV_MENU = "bottomNavigationMenu"
