+++ app/src/main/java/com/android/unio/ui/association/AssociationProfile.kt
@@ -64,7 +64,7 @@ import com.android.unio.model.association.AssociationViewModel
 import com.android.unio.model.association.Member
 import com.android.unio.model.event.Event
 import com.android.unio.model.event.EventViewModel
-import com.android.unio.model.strings.test_tags.AssociationProfileTestTags
+import com.android.unio.model.strings.test_tags.association.AssociationProfileTestTags
 import com.android.unio.model.user.User
 import com.android.unio.model.user.UserViewModel
 import com.android.unio.model.utils.NetworkUtils
